## What type of PR is this? (check all applicable)
- [x] Optimization

## Description
Despite having 700k rows in the taggings table, bc the rows are super small the alter column seems to run pretty damn fast in about 5 seconds. To help mitigate any errors this might cause I plan to increase the postgres STATEMENT_TIMEOUT on the app to 15 seconds so any tagging requests can easily wait for the table to finish updating and still complete their request.

[Here's the timing data](https://gist.github.com/mstruve/1600b29a30b39b9f4edc0c8d79d9ba91)
```
# 5 seconds
"taggings"
# -- change_column(:taggings, :id, :bigint)
#    -> 4.4681s
# irb(main):032:1* Benchmark.realtime {
# irb(main):033:2*   ActiveRecord::Base.connection.execute(
# irb(main):034:0"     <<-SQL
# irb(main):035:0"     ALTER TABLE taggings
# irb(main):036:0"     ALTER COLUMN tag_id TYPE bigint,
# irb(main):037:0"     ALTER COLUMN taggable_id TYPE bigint,
# irb(main):038:0"     ALTER COLUMN tagger_id TYPE bigint
# irb(main):039:2*     SQL
# irb(main):040:1*     )
# irb(main):041:0> }
# => 4.545923504047096

"tag_adjustments"
# -- change_column(:tag_adjustments, :article_id, :bigint)
#    -> 2.4877s
# => []
# irb(main):054:1* Benchmark.realtime {
# irb(main):055:2*   ActiveRecord::Base.connection.execute(
# irb(main):056:0"     <<-SQL
# irb(main):057:0"     ALTER TABLE tag_adjustments
# irb(main):058:0"     ALTER COLUMN user_id TYPE bigint,
# irb(main):059:0"     ALTER COLUMN tag_id TYPE bigint
# irb(main):060:2*     SQL
# irb(main):061:1*     )
# irb(main):062:0> }
# => 0.20964016095967963

"tags"
# -- change_column(:tags, :id, :bigint)
#    -> 0.4116s
# => []
# irb(main):076:1* Benchmark.realtime {
# irb(main):077:2*   ActiveRecord::Base.connection.execute(
# irb(main):078:0"     <<-SQL
# irb(main):079:0"     ALTER TABLE tags
# irb(main):080:0"     ALTER COLUMN badge_id TYPE bigint,
# irb(main):081:0"     ALTER COLUMN mod_chat_channel_id TYPE bigint
# irb(main):082:2*     SQL
# irb(main):083:1*     )
# irb(main):084:0> }
# => 0.3834577740053646
```
## Related Tickets & Documents
https://github.com/forem/forem/projects/9

## Predeploy 
Increase STATEMENT_TIMEOUT to 15 seconds.

![alt_text](https://media1.tenor.com/images/2bfe0b236522489d3fb48666fe48257a/tenor.gif?itemid=5446806)